Issue done but missing unit tests.
